test(Web-UI): Cleanup test files for clarity and consistency
