+++ src/AppHost/ClientApp/tests/nuxt/stores/auth-store/setup.test.ts
