+++ src/AppHost/ClientApp/tests/nuxt/stores/integration-store/integration.test.ts
