+++ src/AppHost/ClientApp/tests/nuxt/stores/server-connection-store/setup.test.ts
