+++ src/AppHost/ClientApp/tests/nuxt/stores/settings-store/setup.test.ts
@@ -1,7 +1,7 @@
 import { describe, beforeAll, beforeEach, test, expect } from 'vitest';
 import { createPinia, setActivePinia } from 'pinia';
 import { subscribeSpyTo, baseSetup, baseVars, getAxiosMock } from '@services-test-base';
-import { generateResultDTO, generateSettingsModel } from '@mock';
+import { generateFailedResultDTO, generateResultDTO, generateSettingsModel } from '@mock';
 import { StoreNames, type ISetupResult } from '@interfaces';
 import { SettingsPaths } from '@api-urls';
 import { useSettingsStore } from '@store';
@@ -36,5 +36,37 @@ describe('SettingsStore.setup()', () => {
 		// Assert
 		expect(result.getFirstValue()).toEqual(setupResult);
 		expect(result.receivedComplete()).toEqual(true);
+	});
+
+	test('Should report failure when refreshing settings fails', async () => {
+		// Arrange
+		const settingsStore = useSettingsStore();
+		mock.onGet(SettingsPaths.getUserSettingsEndpoint()).reply(500, generateFailedResultDTO());
+		const setupResult: ISetupResult = {
+			isSuccess: false,
+			name: StoreNames.SettingsStore,
+		};
+
+		// Act
+		const result = subscribeSpyTo(settingsStore.setup());
+		await result.onComplete();
+
+		// Assert
+		expect(result.getFirstValue()).toEqual(setupResult);
+	});
+
+	test('Should not initialize autosave when refreshing settings fails', async () => {
+		// Arrange
+		const settingsStore = useSettingsStore();
+		mock.onGet(SettingsPaths.getUserSettingsEndpoint()).reply(500, generateFailedResultDTO());
+		mock.onPut(SettingsPaths.updateUserSettingsEndpoint()).reply(200, generateFailedResultDTO());
+		await subscribeSpyTo(settingsStore.setup()).onComplete();
+
+		// Act
+		settingsStore.generalSettings.firstTimeSetup = false;
+		await vi.advanceTimersByTimeAsync(600);
+
+		// Assert
+		expect(mock.history.put.filter((request) => request.url === SettingsPaths.updateUserSettingsEndpoint())).toHaveLength(0);
 	});
 });
